Clean up cost difference tests
