+++ test-loop-tests/src/tests/mod.rs
@@ -1,4 +1,4 @@
-mod account_cost_increase;
+mod account_cost_increase_diff;
 mod backfill_receipt_to_tx;
 mod bandwidth_scheduler;
 #[cfg(feature = "test_features")]
